Start moving to MVVM.
