+++ app/build.gradle
@@ -22,6 +22,9 @@ android {
     androidExtensions {
         experimental = true
     }
+    dataBinding {
+        enabled = true
+    }
 }
 
 dependencies {
