+++ app/src/main/java/ru/ifmo/ctddev/gromov/vkphotosx/model/Api.kt
@@ -1,8 +1,9 @@
-package ru.ifmo.ctddev.gromov.vkphotosx
+package ru.ifmo.ctddev.gromov.vkphotosx.model
 
 import kotlinx.coroutines.Deferred
 import retrofit2.http.GET
 import retrofit2.http.Query
+import ru.ifmo.ctddev.gromov.vkphotosx.model.json.Data
 
 interface Api {
     @GET("method/photos.search")
