+++ app/src/main/java/ru/ifmo/ctddev/gromov/vkphotosx/model/json/Data.kt
@@ -1,0 +1,3 @@
+package ru.ifmo.ctddev.gromov.vkphotosx.model.json
+
+data class Data(var response: Response)
