+++ app/src/main/java/ru/ifmo/ctddev/gromov/vkphotosx/model/json/Item.kt
@@ -1,0 +1,3 @@
+package ru.ifmo.ctddev.gromov.vkphotosx.model.json
+
+data class Item(var id: Int, var text: String, var sizes: List<Size>)
