+++ app/src/main/java/ru/ifmo/ctddev/gromov/vkphotosx/model/json/Size.kt
@@ -1,0 +1,3 @@
+package ru.ifmo.ctddev.gromov.vkphotosx.model.json
+
+data class Size(var url: String)
