+++ app/src/main/java/ru/ifmo/ctddev/gromov/vkphotosx/model/room/AppDB.kt
@@ -1,8 +1,7 @@
-package ru.ifmo.ctddev.gromov.vkphotosx.room
+package ru.ifmo.ctddev.gromov.vkphotosx.model.room
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import ru.ifmo.ctddev.gromov.vkphotosx.Post
 
 @Database(entities = [Post::class], version = 1)
 abstract class AppDB : RoomDatabase() {
