+++ app/src/main/java/ru/ifmo/ctddev/gromov/vkphotosx/view/MainActivity.kt
@@ -1,10 +1,14 @@
-package ru.ifmo.ctddev.gromov.vkphotosx
+package ru.ifmo.ctddev.gromov.vkphotosx.view
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.Observer
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import kotlinx.android.synthetic.main.activity_main.*
+import ru.ifmo.ctddev.gromov.vkphotosx.model.App
+import ru.ifmo.ctddev.gromov.vkphotosx.R
+import ru.ifmo.ctddev.gromov.vkphotosx.viewmodel.RecyclerAdapter
+import ru.ifmo.ctddev.gromov.vkphotosx.model.room.Post
 
 class MainActivity : AppCompatActivity() {
     private lateinit var data: LiveData<List<Post>>
