+++ app/src/main/res/layout/post.xml
@@ -1,9 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
-<TextView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/post"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="10dp"
-    android:layout_marginBottom="10dp"
-    android:gravity="center_horizontal"
-    android:padding="10dp" />
+<layout>
+
+    <data>
+
+        <variable
+            name="post"
+            type="ru.ifmo.ctddev.gromov.vkphotosx.model.room.Post" />
+    </data>
+
+    <TextView xmlns:android="http://schemas.android.com/apk/res/android"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="10dp"
+        android:layout_marginBottom="10dp"
+        android:background="@{post.isFave ? @color/fave_post : @color/post}"
+        android:gravity="center_horizontal"
+        android:onClickListener="@{() -> ru.ifmo.ctddev.gromov.vkphotosx.model.room.PostKt.onClick(post, context)}"
+        android:padding="10dp"
+        android:text="@={post.text}" />
+</layout>
